requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
attrs==24.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via streamlit
dnspython==2.7.0
    # via pymongo
ecdsa==0.19.0
    # via python-jose
fastapi==0.115.6
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
greenlet==3.1.1
    # via playwright
idna==3.10
    # via
    #   anyio
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.5
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.21.1
    # via altair
numpy==2.2.1
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   pytest
    #   streamlit
pandas==2.2.3
    # via streamlit
passlib==1.7.4
    # via -r requirements.in
pillow==11.1.0
    # via
    #   -r requirements.in
    #   streamlit
playwright==1.49.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
protobuf==5.29.2
    # via streamlit
pyarrow==18.1.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pydantic==2.10.5
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pyee==12.0.0
    # via playwright
pygments==2.19.1
    # via rich
pymongo==4.10.1
    # via -r requirements.in
pytest==8.3.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-jose==3.3.0
    # via -r requirements.in
pytz==2024.2
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   -r requirements.in
    #   streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via -r requirements.in
starlette==0.41.3
    # via fastapi
streamlit==1.41.1
    # via -r requirements.in
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pyee
    #   sqlalchemy
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via requests
